refactor(ui): decouple AI query from DSLFilterConditionField via composition
The base field is now a pure DSL input with a generic composition
surface: a prose variant, injected CodeMirror extensions, leading-visual
and extra-controls slots, focus/clear callbacks, and an imperative focus
ref. All AI query behavior (mode toggle, conversion phase machine,
keymap, badges, outline, prevalidation) moves to AIQueryDSLFilterField
in ai/, built entirely on that surface.

Also adds Mod-Enter to hand a DSL-mode draft to AI query (with a hint
badge while typing) and disables Shift-Enter outright in the
single-line field.
