+++ app/src/components/filter/styles.ts
@@ -1,6 +1,5 @@
 import { css, keyframes } from "@emotion/react";
 
-import { svgSize as pxiGlyphSize } from "@phoenix/components/agent/PxiGlyph";
 import { NON_MODAL_FLOATING_Z_INDEX } from "@phoenix/components/core/zIndex";
 
 /**
@@ -138,48 +137,17 @@ export const dslFilterErrorTooltipCSS = css`
 `;
 
 /**
- * Grows the error badge out from the editor's right edge. Animating
+ * Grows a control-cluster badge out from the editor's right edge. Animating
  * max-width alongside opacity keeps the appearance smooth — the editor
  * cedes the space gradually instead of the badge popping in at full size.
+ * Exported so composed badges (e.g. the AI-query ones) grow in the same way.
  */
-const errorBadgeIn = keyframes`
+export const dslFilterBadgeGrowIn = keyframes`
   from {
     opacity: 0;
     max-width: 0;
     padding-left: 0;
     padding-right: 0;
-  }
-`;
-
-/**
- * Sizes the PxiOutline wrapper like the bare field used to size itself and
- * silences the outline's resting stroke — the field should look untouched
- * until AI query actually engages (a natural-language draft or an
- * in-flight conversion).
- */
-export const dslFilterAIOutlineCSS = css`
-  flex: 1 1 auto;
-  min-width: 0;
-  /* The gradient hugs the field's border — the outline's default standoff
-     gap reads as a detached double border on an input */
-  --pxi-outline-gap: 0px;
-  &[data-state="idle"] .pxi-outline__stroke {
-    opacity: 0;
-  }
-  /* While the generative border is showing it is the focus affordance —
-     focus brings the band to full strength (the field suppresses its own
-     theme focus ring, which would clash with the gradient) */
-  &:not([data-state="idle"]):has(.cm-content:focus-visible)
-    .pxi-outline__stroke {
-    opacity: 1;
-  }
-  /* The outline isolates its stacking (for the glow layers), which traps
-     the typeahead's own z-index inside it — later-stacked page content
-     like a table's sticky header would paint over the open dropdown.
-     Elevate the whole outline only while a tooltip is showing so the
-     field doesn't sit above sibling floating UI the rest of the time. */
-  &:has(.cm-tooltip) {
-    z-index: ${NON_MODAL_FLOATING_Z_INDEX};
   }
 `;
 
@@ -202,18 +170,6 @@ export const dslFilterFieldCSS = css`
     outline: var(--focus-ring-thickness) solid var(--focus-ring-color);
     outline-offset: calc(-1 * var(--focus-ring-thickness));
   }
-  /* In plain-English mode the generative gradient border carries focus
-     (brightening to full strength) — the theme ring on top of it reads
-     as two clashing borders */
-  &[data-ai-mode="true"]:has(.cm-content:focus-visible) {
-    outline: none;
-  }
-  /* When the gradient wraps the field it IS the border — the input's own
-     border under it reads as a second, internal ring. Placed after the
-     hover/focus border rules so it wins them by source order. */
-  &[data-ai-mode="true"] {
-    border-color: transparent;
-  }
   /* Flag invalidity only once the user has left the field — a red border
      while they're still typing/fixing the expression is too alarming */
   &[data-is-invalid="true"]:not([data-is-focused="true"]) {
@@ -223,20 +179,6 @@ export const dslFilterFieldCSS = css`
   .filter-icon {
     margin-left: var(--global-dimension-size-100);
     margin-right: var(--global-dimension-size-50);
-  }
-  /* The converting glyph keeps the resting glyph's footprint (no layout
-     shift) and wears the same thinking tint as the Ask PXI nav button */
-  .dsl-filter-condition-field__thinking-glyph {
-    display: grid;
-    place-items: center;
-    flex: none;
-    width: ${pxiGlyphSize}px;
-    height: ${pxiGlyphSize}px;
-    color: color-mix(
-      in srgb,
-      var(--pxi-treatment-color-middle) 78%,
-      var(--pxi-treatment-color-end)
-    );
   }
   /* Everything after the editor — badges, the mode toggle, settings, and
      clear — shares one flex group so spacing comes from a single gap
@@ -262,7 +204,7 @@ export const dslFilterFieldCSS = css`
     line-height: var(--global-line-height-xs);
     white-space: nowrap;
     cursor: default;
-    animation: ${errorBadgeIn} 0.25s ease-out;
+    animation: ${dslFilterBadgeGrowIn} 0.25s ease-out;
     @media (prefers-reduced-motion: reduce) {
       animation: none;
     }
@@ -291,78 +233,14 @@ export const dslFilterFieldCSS = css`
        interpolable: a fixed max and a free min */
     min-width: 0;
     max-width: var(--global-dimension-size-250);
-    animation: ${errorBadgeIn} 0.25s ease-out;
+    animation: ${dslFilterBadgeGrowIn} 0.25s ease-out;
     @media (prefers-reduced-motion: reduce) {
       animation: none;
     }
   }
-  /* Plain-English mode reads as prose, not code — mirror that in the
-     editor; the leading PXI glyph (self-colored, subtly animated) marks
-     the mode at a glance even while the field is empty */
-  &[data-ai-mode="true"] .cm-content {
+  /* The prose variant reads as prose, not code — mirror that in the
+     editor's font */
+  &[data-variant="prose"] .cm-content {
     font-family: var(--global-font-family-sans);
-  }
-  /* One treatment spec for the engaged AI controls — the badge and the
-     pressed sparkle toggle read as one family, while the neutral buttons
-     (gear, clear) stay gray */
-  .ai-badge,
-  .ai-mode-toggle[aria-pressed="true"] {
-    color: var(--pxi-treatment-color-middle);
-    border-color: color-mix(
-      in srgb,
-      var(--pxi-treatment-color-middle) 35%,
-      transparent
-    );
-    background-color: color-mix(
-      in srgb,
-      var(--pxi-treatment-color-middle) 12%,
-      transparent
-    );
-  }
-  .ai-badge {
-    display: flex;
-    align-items: center;
-    flex-shrink: 0;
-    gap: var(--global-dimension-size-50);
-    padding: 2px var(--global-dimension-size-65);
-    border-radius: var(--global-rounding-small);
-    border-width: var(--global-border-size-thin);
-    border-style: solid;
-    font-size: var(--global-font-size-xs);
-    line-height: var(--global-line-height-xs);
-    white-space: nowrap;
-    cursor: default;
-    animation: ${errorBadgeIn} 0.25s ease-out;
-    @media (prefers-reduced-motion: reduce) {
-      animation: none;
-    }
-  }
-  .ai-undo-button {
-    display: flex;
-    align-items: center;
-    gap: var(--global-dimension-size-25);
-    padding: 0 var(--global-dimension-size-50);
-    border-radius: var(--global-rounding-small);
-    color: inherit;
-    font-size: var(--global-font-size-xs);
-    cursor: pointer;
-    &:hover {
-      background-color: color-mix(
-        in srgb,
-        var(--pxi-treatment-color-middle) 18%,
-        transparent
-      );
-    }
-    &:focus-visible {
-      outline: var(--focus-ring-thickness) solid var(--focus-ring-color);
-      outline-offset: var(--focus-ring-offset);
-    }
-  }
-  .ai-mode-toggle[aria-pressed="true"]:hover {
-    background-color: color-mix(
-      in srgb,
-      var(--pxi-treatment-color-middle) 20%,
-      transparent
-    );
   }
 `;
